Always show the panel on a primary output.
Previously width and height were taken using Xrandr functions
for (presumably) primary output. But actual rendering didn't account
for the fact that primary output can cover arbitrary part of abstract
X11 screen. For example, I usually have my laptop's screen put to the
right of external screen. Thus the previous code would render login
screen on the external screen but using dimensions of laptops
screen (because it's set to be primary).

So the fix is to get primary output's viewport in the abstract screen
and perform rendering to it.